blog handler refactored
